Reorganise the render/ package
