+++ render/host_test.go
@@ -1,0 +1,19 @@
+package render_test
+
+import (
+	"testing"
+
+	"github.com/weaveworks/scope/render"
+	"github.com/weaveworks/scope/render/expected"
+	"github.com/weaveworks/scope/test"
+	"github.com/weaveworks/scope/test/fixture"
+	"github.com/weaveworks/scope/test/reflect"
+)
+
+func TestHostRenderer(t *testing.T) {
+	have := render.HostRenderer.Render(fixture.Report).Prune()
+	want := expected.RenderedHosts.Prune()
+	if !reflect.DeepEqual(want, have) {
+		t.Error(test.Diff(want, have))
+	}
+}
