+++ render/id.go
@@ -2,6 +2,8 @@ package render
 
 import (
 	"strings"
+
+	"github.com/weaveworks/scope/report"
 )
 
 // MakePseudoNodeID joins the parts of an id into the id of a pseudonode
@@ -12,4 +14,14 @@ func MakePseudoNodeID(parts ...string) string {
 // MakeGroupNodeTopology joins the parts of a group topology into the topology of a group node
 func MakeGroupNodeTopology(originalTopology, key string) string {
 	return strings.Join([]string{"group", originalTopology, key}, ":")
+}
+
+// NewDerivedNode makes a node based on node, but with a new ID
+func NewDerivedNode(id string, node report.Node) report.Node {
+	return node.WithID(id).WithChildren(report.MakeNodeSet(node)).PruneParents()
+}
+
+// NewDerivedPseudoNode makes a new pseudo node with the node as a child
+func NewDerivedPseudoNode(id string, node report.Node) report.Node {
+	return NewDerivedNode(id, node).WithTopology(Pseudo)
 }
